requirements/production.txt
# PRECAUTION: avoid production dependencies that aren't in development

-r base.txt

gunicorn==20.0.4  # https://github.com/benoitc/gunicorn
psycopg2-binary==2.8.6  # https://github.com/psycopg/psycopg2
sentry-sdk==0.19.5  # https://github.com/getsentry/sentry-python

# Django
# ------------------------------------------------------------------------------
django-anymail==8.1  # https://github.com/anymail/django-anymail
django-annoying==0.10.6
django-crontab==0.7.1
django_extensions==3.1.0

# data manipulation libraries
nltk[all]
pandas==1.2.0
wordcloud==1.8.1